+++ constants/Colors.js
@@ -20,4 +20,5 @@ export default {
   indicatorLime: "#A6BC35",
   indicatorGreen: "#6DD149",
   linkBlue: "#2980b9",
+  tabBackground: "#ced0d9",
 };
